move findSupplier to search controller
