+++ app/pages/sections/supplier/search/controller.js
@@ -1,3 +1,6 @@
+import { getData } from 'buying-catalogue-library';
+import { getEndpoint } from '../../../../endpoints';
+import { logger } from '../../../../logger';
 import { getContext, getErrorContext } from './contextCreator';
 
 export const getSupplierSearchPageContext = async ({ orderId }) => (
@@ -16,6 +19,14 @@ export const validateSupplierSearchForm = ({ data }) => {
     },
   ];
   return { success: false, errors };
+};
+
+export const findSuppliers = async ({ name, accessToken }) => {
+  const endpoint = getEndpoint({ endpointLocator: 'getSearchSuppliers', options: { name } });
+  const suppliersFound = await getData({ endpoint, accessToken, logger });
+  logger.info(`Searching for "${name}" returned ${suppliersFound.length} supplier`);
+
+  return suppliersFound;
 };
 
 export const getSupplierSearchPageErrorContext = params => getErrorContext(params);
