+++ app/pages/sections/supplier/search/controller.test.js
@@ -1,5 +1,10 @@
+import { getData } from 'buying-catalogue-library';
+import { solutionsApiUrl } from '../../../../config';
+import { logger } from '../../../../logger';
 import * as contextCreator from './contextCreator';
-import { getSupplierSearchPageContext, validateSupplierSearchForm } from './controller';
+import {
+  getSupplierSearchPageContext, validateSupplierSearchForm, findSuppliers,
+} from './controller';
 
 jest.mock('buying-catalogue-library');
 
@@ -68,6 +73,25 @@ describe('supplier search controller', () => {
         const response = validateSupplierSearchForm({ data });
 
         expect(response.errors).toEqual(expectedValidationErrors);
+      });
+    });
+  });
+
+  describe('findSuppliers', () => {
+    afterEach(() => {
+      getData.mockReset();
+    });
+
+    it('should call getData once with the correct params', async () => {
+      getData
+        .mockResolvedValueOnce({ data: [] });
+
+      await findSuppliers({ name: 'some-supp', accessToken: 'access_token' });
+      expect(getData.mock.calls.length).toEqual(1);
+      expect(getData).toHaveBeenCalledWith({
+        endpoint: `${solutionsApiUrl}/api/v1/suppliers?name=some-supp`,
+        accessToken: 'access_token',
+        logger,
       });
     });
   });
